src/functions/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
pydantic
azure-storage-blob
aiohttp
azure-identity
azure-ai-documentintelligence